% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{format_data}
\alias{format_data}
\title{Apply formatting to data.frame content}
\usage{
format_data(d, colFormats = NULL, typeFormats = NULL, useOptions = TRUE,
  na.rm = getOption("tblr.na.rm", TRUE))
}
\description{
Apply formatting to data.frame content
}

