% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{HTMLencode}
\alias{HTMLencode}
\title{Convert certain characters to HTML encoding}
\usage{
HTMLencode(s)
}
\description{
Convert certain characters to HTML encoding, eg. '<' becomes '&amp;'
Special symbols can be specified as LaTeX (TODO)
}
\references{
\url{http://www.escapecodes.info}
}

