% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{rtable}
\alias{rtable}
\alias{rtable-package}
\title{High-level specification of table formatting for use with knitr/Sweave}
\description{
High-level specification of table formatting for use with knitr/Sweave
}

