% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tbl.data.frame}
\alias{tbl.data.frame}
\title{Create a table object from a data.frame}
\usage{
\method{tbl}{data.frame}(d, colFormats = NULL, typeFormats = NULL,
  useOptions = TRUE, na.rm = getOption("tbl.na.rm", TRUE), row.names = T,
  col.names = T, corner = "", ...)
}
\description{
Create a table object from a data.frame
}

