% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R, R/tblr-package.R
\docType{package}
\name{tblr}
\alias{tblr}
\alias{tblr-package}
\title{High-level specification of table formatting for use with knitr/Sweave}
\description{
High-level specification of table formatting for use with knitr/Sweave

tblr.
}

