% Generated by roxygen2 (4.0.2): do not edit by hand
\name{format_data}
\alias{format_data}
\title{Apply formatting to data.frame content}
\usage{
format_data(d, colFormats = NULL, typeFormats = NULL, useOptions = TRUE,
  na.rm = getOption("tbl.na.rm", TRUE))
}
\description{
Apply formatting to data.frame content
}

